Ticket #1247: Fix size type itf

+++ samples/ClientServer/demo_push_server/push_server_lib/internal/sopc_push_server_config.c
@@ -317,7 +317,7 @@ SOPC_ReturnStatus SOPC_PushServerConfig_GetDefaultConfiguration(SOPC_PKIProvider
                                                                 SOPC_SerializedCertificate* pServerCert,
                                                                 SOPC_PKIProvider* pPKIUsr,
                                                                 const SOPC_Certificate_Type usrCertType,
-                                                                const size_t maxTrustListSize,
+                                                                const uint32_t maxTrustListSize,
                                                                 SOPC_PushServerConfig_Config** ppConfig)
 {
     if (NULL == pPKIApp || 0 == maxTrustListSize || NULL == pServerKey || NULL == pServerCert || NULL == ppConfig)
